+++ wandb/sdk/data_types/helper_types/__init__.py
@@ -1,0 +1,9 @@
+from .BoundingBoxes2D import BoundingBoxes2D
+from .Classes import Classes
+from .ImageMask import ImageMask
+
+__all__ = [
+    "BoundingBoxes2D",
+    "Classes",
+    "ImageMask",
+]
